Move tests to bridge tests
