+++ src/System.Diagnostics.DiagnosticSource/tests/DiagnosticSourceTests.cs
@@ -685,33 +685,6 @@ namespace System.Diagnostics.Tests
             }
         }
 
-        
-        [Fact]
-        public void IndexGetters_DontThrow()
-        {
-            using (var eventListener = new TestDiagnosticSourceEventListener())
-            using (var diagnosticListener = new DiagnosticListener("MySource"))
-            {
-                eventListener.Enable(
-                    "MySource/MyEvent"
-                );
-                diagnosticListener.Write(
-                    "MyEvent",
-                    new MyEvent
-                    {
-                        Number = 1,
-                        OtherNumber = 2
-                    }
-                );
-                Assert.Equal(1, eventListener.EventCount);
-                Assert.Equal("MySource", eventListener.LastEvent.SourceName);
-                Assert.Equal("MyEvent", eventListener.LastEvent.EventName);
-                Assert.True(2 <= eventListener.LastEvent.Arguments.Count);
-                Assert.Equal("1", eventListener.LastEvent.Arguments["Number"]);
-                Assert.Equal("2", eventListener.LastEvent.Arguments["OtherNumber"]);
-            }
-        }
-
         #region Helpers
         /// <summary>
         /// Returns the list of active diagnostic listeners.
@@ -811,21 +784,5 @@ namespace System.Diagnostics.Tests
     {
         public string Name { get; set; }
         public int Id { get; set; }
-    }
-
-    /// <summary>
-    /// classes for test data
-    /// </summary>
-    internal class MyEvent
-    {
-        public int Number { get; set; }
-        public int OtherNumber { get; set; }
-        public int Count => 2;
-        public KeyValuePair<string, object> this[int index] => index switch
-        {
-            0 => new KeyValuePair<string, object>(nameof(Number), Number),
-            1 => new KeyValuePair<string, object>(nameof(OtherNumber), OtherNumber),
-            _ => throw new IndexOutOfRangeException()
-        };
     }
 }
